// src/types/interfaces/profile.interface.ts
import { Types } from 'mongoose';

import { UserStatus } from '../constants/user-status.constants';
import { Role } from '../enums/role.enums';
import { IProfileDocumentFile } from './profile-document.interface';
import { IMission, IQuarter } from './statement.interface';
export interface IProfile {
  _id: string | null;
  userId: string | null;
  email: string;
  firstName: string;
  lastName: string;
  fullName: string;
  contactNumber?: string;
  employeeId?: string;
  status: string | null;
  profilePicture?: IProfileDocumentFile;
  resume?: IProfileDocumentFile;
  idProof?: IProfileDocumentFile;
  documents: IProfileDocumentFile[];
  dateOfBirth?: string | null;
  designation?: string;
  role: string | null;
  salary?: string;
  emergencyContact?: string;
  currentAddress?: string;
  permanentAddress?: string;
  dateOfJoining?: string | null;
  leftUs?: string | null;
  rejoinUs?: string | null;
  skills: string[];
  achievements: string[];
  missionStatement?: IMission | null;
  successIndicators?: IQuarter[] | null;
  createdBy?: string | null;
  updatedBy?: string | null;
  createdAt?: Date;
  updatedAt?: Date;
}
export interface ProfileData {
  _id: Types.ObjectId;
  userId: Types.ObjectId;
  fullName: string;
  email: string;
  firstName: string;
  lastName: string;
  status: UserStatus;
  resume?: IProfileDocumentFile;
  idProof?: IProfileDocumentFile;
  documents?: IProfileDocumentFile[];
  dateOfBirth?: Date;
  designation?: string;
  role: Role;
  employeeId?: string;
  salary?: string;
  contactNumber?: string;
  emergencyContact?: string;
  currentAddress?: string;
  permanentAddress?: string;
  profilePicture?: IProfileDocumentFile;
  dateOfJoining?: Date;
  leftUs?: Date;
  rejoinUs?: Date;
  skills?: string[];
  achievements?: string[];
  missionStatement?: IMission | null;
  successIndicators?: IQuarter[] | null;
  createdBy?: Types.ObjectId | null;
  updatedBy?: Types.ObjectId | null;
  createdAt?: Date;
  updatedAt?: Date;
}
export interface TransformedProfile {
  employeeId?: string;
  _id: string;
  userId: string;
  fullName: string;
  email: string;
  firstName: string;
  lastName: string;
  designation?: string;
  role: string;
  salary: string;
  status: UserStatus;
  dateOfBirth?: string;
  dateOfJoining?: string;
  contactNumber: string;
  emergencyContact: string;
  currentAddress: string;
  permanentAddress: string;
  resume?: IProfileDocumentFile;
  idProof?: IProfileDocumentFile;
  profilePicture?: IProfileDocumentFile;
  skills: string[];
  achievements: string[];
  documents: IProfileDocumentFile[];
  missionStatement?: IMission | null;
  successIndicators?: IQuarter[] | null;
}